library-backend/src/main/java/ru/dankoy/library/core/service/genre/GenreServiceMongo.java
package ru.dankoy.library.core.service.genre;

import java.util.List;
import java.util.Set;
import java.util.stream.Collectors;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import ru.dankoy.library.core.domain.Genre;
import ru.dankoy.library.core.domain.Work;
import ru.dankoy.library.core.service.work.WorkService;

@Service
@RequiredArgsConstructor
public class GenreServiceMongo implements GenreService {

  private final WorkService workService;

  @Override
  public void update(Genre oldGenre, Genre newGenre) {

    List<Work> works = workService.findAllByGenreName(oldGenre);

    works.forEach(
        b -> {
          b.getGenres().remove(oldGenre);
          b.getGenres().add(newGenre);
        });

    workService.updateMultiple(works);
  }

  @Override
  public void delete(Genre genre) {

    List<Work> works = workService.findAllByGenreName(genre);

    works.forEach(b -> b.getGenres().remove(genre));

    workService.updateMultiple(works);
  }

  @Override
  public Set<Genre> getAllGenres() {
    var books = workService.findAll();

    return books.stream().flatMap(b -> b.getGenres().stream()).collect(Collectors.toSet());
  }
}
